brandgenie-pro-web/backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.1.0
SQLAlchemy==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10
httpx==0.28.1
requests==2.32.4
urllib3==2.5.0
charset_normalizer==3.4.2
bcrypt==4.3.0
passlib==1.7.4
ecdsa==0.19.1
pyasn1==0.6.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
dnspython==2.7.0
email-validator==2.2.0
uuid==1.30
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
oauthlib==3.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1-modules==0.4.2
pyparsing==3.2.3
requests-oauthlib==2.0.0
uritemplate==4.2.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
h11==0.16.0
httpcore==1.0.9
idna==3.10
greenlet==3.2.3
pydantic==2.11.7
pydantic-core==2.33.2
starlette==0.46.2
typing-extensions==4.14.0
typing-inspection==0.4.1
annotated-types==0.7.0
sniffio==1.3.1
colorama==0.4.6
python-multipart==0.0.20
pytz==2025.2
